- Resolve merge conflicts: adopt struct-based instruction args
  (CreateWalletArgs, ProposeArgs, IntentInner) from upstream while
  preserving dWallet fields (chain_kind, tx_template_offset/len)
- Update vendored quasar_client: DynVec for proposers/approvers,
  add proposal_index to propose
- Update ika CPI for new pre-alpha SDK: add coordinator account to
  approve_message, message_metadata_digest param, signature_scheme
  u8->u16, hierarchical MessageApproval PDA seeds
- Update CLI ika.rs: versioned DKG/presign attestations,
  DWalletSignatureScheme combined enum, new Sign request format,
  dWallet PDA curve u16, updated MA layout offsets

+++ cli/src/quasar_client/bind_dwallet.rs
@@ -14,7 +14,7 @@ pub struct BindDwalletInstruction {
     pub system_program: Address,
     pub chain_kind: u8,
     pub user_pubkey: [u8; 32],
-    pub signature_scheme: u8,
+    pub signature_scheme: u16,
     pub cpi_authority_bump: u8,
 }
